requirements.txt
requests==2.4.3
PyGithub==1.25.2
